refact: rename config files
